requirements.txt
fastapi>=0.104.1
pydantic>=2.4.2
spacy>=3.7.2
uvicorn>=0.23.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
dnspython>=2.4.2
pydantic[email]
